requirements.txt
numpy==1.26.4
Pillow==10.3.0
scipy==1.11.4
wxPython==4.2.1
